examples/pd_grid: Cleanup and refactor
- Flatten structure
- Completely rewrite SolaraViz
- Cleanup and update docs
