--- examples/advanced/pd_grid/requirements.txt
@@ -1,3 +1,0 @@
-jupyter
-matplotlib
-mesa~=2.0
