--- examples/advanced/pd_grid/run.py
@@ -1,3 +1,0 @@
-from pd_grid.server import server
-
-server.launch(open_browser=True)
